backend/pyproject.toml
[tool.poetry]
name = "elfai"
version = "1.0.0"
description = "Elf AI API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "apps"}, {include = "config"}]
license = "MIT"
homepage = "https://github.com/yourusername/elfai"
repository = "https://github.com/yourusername/elfai"
keywords = ["media", "automation", "django", "transmission"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Video",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
django-health-check = "^3.18.3"
django-redis = "^5.4.0"
django-structlog = "^7.0.0"
psutil = "^5.9.8"
psycopg = {extras = ["binary"], version = "^3.1.18"}
redis = "^5.0.1"
structlog = "^24.1.0"
whitenoise = "^6.6.0"
django-filter = "^23.5"
drf-spectacular = "^0.27.0"
pydantic = {extras = ["email"], version = "^2.6.0"}
pydantic-settings = "^2.1.0"
transmission-rpc = "^7.0.3"
requests = "^2.31.0"
beautifulsoup4 = {extras = ["lxml"], version = "^4.12.0"}
python-json-logger = "^2.0.7"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-django = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
sentry-sdk = {extras = ["django"], version = "^1.40.0"}
gunicorn = "^21.2.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
django-silk = "^5.1.0"
tqdm = "^4.66.1"
celery = "^5.3.6"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.3.0"
openai = "^1.12.0"
prometheus_client = "^0.20.0"
mysqlclient = "^2.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-env = "^1.1.3"
pytest-asyncio = "^0.23.3"
pytest-benchmark = "^4.0.0"
hypothesis = "^6.98.0"
schemathesis = "^3.25.0"
mutmut = "^2.4.3"
black = "^24.1.1"
ruff = "^0.1.14"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-docstrings = "^1.7.0"
flake8-bugbear = "^24.2.6"
mypy = "^1.8.0"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
types-requests = "^2.31.0.20240125"
factory-boy = "^3.3.0"
bandit = "^1.7.7"
safety = "^2.3.5"
vulture = "^2.10"
pyupgrade = "3.15.0"
pre-commit = "^3.6.0"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
sphinx-autodoc-typehints = "^1.25.2"
myst-parser = "^2.0.0"
debugpy = "^1.8.0"

[tool.poetry.scripts]
elfai = "cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["apps", "config"]
known_django = ["django", "rest_framework"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main", "pydantic.mypy"]
exclude = [
    '^docs/',
    'migrations/',
    'tests/'
]

[[tool.mypy.overrides]]
module = ["*.migrations.*", "*.tests.*"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.base"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = """
    --doctest-modules
    --cov=apps
    --cov-report=term-missing
    --cov-report=html
    --strict-markers
    --strict-config
    --tb=short
    --benchmark-only
    --benchmark-group-by=func
    --benchmark-disable-gc
    -v
"""
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__init__.py",
    "*/admin.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
fail_under = 80

[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B311"]  # Standard pseudo-random generators are not suitable for security/cryptographic purposes.

[tool.vulture]
exclude = ["migrations", "tests"]
min_confidence = 80
